requirements.txt
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
jsonschema==4.20.0
